infra/pyproject.toml
[project]
authors = [
  {name = "Amichay Oren", email = "amichay@nine30.com"},
]
dependencies = [
  "requests>=2.30.0",
  "flask>=2.3.2",
  "google-cloud>=0.34.0",
  "google-crc32c>=1.5.0",
  "keyrings-google-artifactregistry-auth>=1.1.2",
]
description = "Infrastructure code, shared between several GCP Functions"
license = {text = "MIT"}
name = "infra"
readme = "README.md"
requires-python = ">=3.11"
version = "0.1.45"

[build-system]
build-backend = "pdm.backend"
requires = ["pdm-backend"]
